fix dtype

+++ test/sagemaker_tests/huggingface/inference/resources/code/flash_attn_inference.py
@@ -11,7 +11,7 @@ print("running flash attention test...")
 def model_fn(model_dir):
     tokenizer = AutoTokenizer.from_pretrained(model_dir)
     model = AutoModel.from_pretrained(
-        model_dir, torch_dtype=torch.float32, attn_implementation="flash_attention_2"
+        model_dir, torch_dtype=torch.float16, attn_implementation="flash_attention_2"
     )
     model.to("cuda")
     return {"model": model, "tokenizer": tokenizer}
